Responding to review; seperating files
